Move tablemetadata
